Pedestal substraction is splited into two steps

+++ Analysis/src/PedMeanCalc.cpp
@@ -1,7 +1,7 @@
 #include "PedMeanCalc.hpp"
 
 PedMeanCalc::PedMeanCalc() {
-	eventIter_ = NULL;
+	ped_file_ = NULL;
 	done_flag_ = false;
 	for (int i = 0; i < 25; i++) {
 		for (int j = 0; j < 64; j++) {
@@ -15,72 +15,80 @@ PedMeanCalc::PedMeanCalc() {
 }
 
 PedMeanCalc::~PedMeanCalc() {
-	destroy_();
-}
-
-void PedMeanCalc::create_() {
-	if (eventIter_ == NULL)
-		return;
-	for (int i = 0; i < 25; i++) {
-		for (int j = 0; j < 64; j++) {
-			sprintf(name_, "f_gaus_%d_%d", i + 1, j + 1);
-			f_gaus_[i][j] = new TF1(name_, "gaus(0)", 0, PED_MAX);
-			f_gaus_[i][j]->SetParameters(1, 400, 25);
-			f_gaus_[i][j]->SetParName(0, "Amplitude");
-			f_gaus_[i][j]->SetParName(1, "Mean");
-			f_gaus_[i][j]->SetParName(2, "Sigma");
-			sprintf(name_, "h_ped_%d_%d", i + 1, j + 1);
-			sprintf(title_, "Pedestal of CH %d_%d", i + 1, j + 1);
-			h_ped_[i][j] = new TH1F(name_, title_, PED_BINS, 0, PED_MAX);
-			h_ped_[i][j]->SetDirectory(NULL);
-		}
-	}
-}
-
-void PedMeanCalc::destroy_() {
-	if (eventIter_ != NULL)
-		eventIter_ = NULL;
+	if (ped_file_ != NULL)
+		close();
 	if (h_ped_map_ != NULL) {
 		delete h_ped_map_;
 		h_ped_map_ = NULL;
 	}
+}
+
+bool PedMeanCalc::open(const char* filename, char m) {
+	if (m == 'w' || m == 'W')
+		mode_ = 'w';
+	else if (m == 'r' || m == 'R')
+		mode_ = 'r';
+	else
+		return false;
+	
+	if (mode_ == 'r')
+		ped_file_ = new TFile(filename, "READ");
+	else
+		ped_file_ = new TFile(filename, "UPDATE");
+	if (ped_file_->IsZombie())
+		return false;
+	
 	for (int i = 0; i < 25; i++) {
 		for (int j = 0; j < 64; j++) {
-			if (f_gaus_[i][j] != NULL) {
-				delete f_gaus_[i][j];
-				f_gaus_[i][j] = NULL;
+			if (mode_ == 'r') {
+				sprintf(name_, "f_gaus_%d_%d", i + 1, j + 1);
+				f_gaus_[i][j] = new TF1(name_, "gaus(0)", 0, PED_MAX);
+				f_gaus_[i][j]->SetParameters(1, 400, 25);
+				f_gaus_[i][j]->SetParName(0, "Amplitude");
+				f_gaus_[i][j]->SetParName(1, "Mean");
+				f_gaus_[i][j]->SetParName(2, "Sigma");
 			}
-			if (h_ped_[i][j] != NULL) {
-				delete h_ped_[i][j];
-				h_ped_[i][j] = NULL;
+			sprintf(name_, "h_ped_%d_%d", i + 1, j + 1);
+			h_ped_[i][j] = static_cast<TH1F*>(ped_file_->Get(name_));
+			if (h_ped_[i][j] == NULL) {
+				if (mode_ == 'r') {
+					return false;
+				} else {
+					sprintf(title_, "Pedestal of CH %d_%d", i + 1, j + 1);
+					h_ped_[i][j] = new TH1F(name_, title_, PED_BINS, 0, PED_MAX);
+					h_ped_[i][j]->SetDirectory(ped_file_);
+				}
 			}
 		}
 	}
+	return true;
 }
 
-void PedMeanCalc::init(EventIterator* iter) {
-	destroy_();
-	eventIter_ = iter;
-	done_flag_ = false;	
-	create_();
-}
-
-void PedMeanCalc::do_calc() {
-	if (eventIter_ == NULL)
+void PedMeanCalc::do_fill(EventIterator& eventIter) {
+	if (ped_file_ == NULL)
 		return;
-	eventIter_->ped_trigg_restart();
-	while (eventIter_->ped_trigg_next()) {
-		while (eventIter_->ped_event_next()) {
-			int idx = eventIter_->ped_event.ct_num - 1;
+	if (mode_ != 'w')
+		return;
+	eventIter.ped_trigg_restart();
+	while (eventIter.ped_trigg_next()) {
+		while (eventIter.ped_event_next()) {
+			int idx = eventIter.ped_event.ct_num - 1;
 			for (int j = 0; j < 64; j++) {
-				Int_t ch = eventIter_->ped_event.energy_ch[j];
+				Int_t ch = eventIter.ped_event.energy_ch[j];
 				if (ch > PED_MAX)
 					continue;
 				h_ped_[idx][j]->Fill(static_cast<Float_t>(ch));
 			}
 		}
 	}
-	
+	eventIter.ped_trigg_restart();
+}
+
+void PedMeanCalc::do_fit() {
+	if (ped_file_ == NULL)
+		return;
+	if (mode_ != 'r')
+		return;
 	for (int i = 0; i < 25; i++) {
 		for (int j = 0; j < 64; j++) {
 			mean_0[i][j] = h_ped_[i][j]->GetMean();
@@ -93,27 +101,51 @@ void PedMeanCalc::do_calc() {
 			Float_t min_ch = mean_0[i][j] - 200 > 0 ? mean_0[i][j] - 200 : 0;
 			Float_t max_ch = mean_0[i][j] + 200 < PED_MAX ? mean_0[i][j] + 200 : PED_MAX;
 			f_gaus_[i][j]->SetParameter(1, mean_0[i][j]);
-			f_gaus_[i][j]->SetParameter(2, sigma_0[i][j]);
+			f_gaus_[i][j]->SetParameter(2, sigma_0[i][j] - 1);
 			f_gaus_[i][j]->SetRange(min_ch, max_ch);
+			h_ped_[i][j]->Fit(f_gaus_[i][j], "RQN");
 			h_ped_[i][j]->Fit(f_gaus_[i][j], "RQN");
 			mean[i][j] = f_gaus_[i][j]->GetParameter(1);
 			sigma[i][j] = abs(f_gaus_[i][j]->GetParameter(2));
 			if (mean[i][j] < 0 || mean[i][j] > PED_MAX) {
-				cout << "- Warning: mean of CH " << i + 1 << "_" << j + 1 << " is out of range, use the arithmetic mean instead." << endl;
+				cout << "- Warning: mean of CH " << i + 1 << "_" << j + 1
+					 << " is out of range, use the arithmetic mean instead." << endl;
 				mean[i][j] = mean_0[i][j];
 				sigma[i][j] = sigma_0[i][j];
 			} else if (sigma[i][j] / sigma_0[i][j] > 2.0) {
-				cout << "- Warning: sigma of CH " << i + 1 << "_" << j + 1 << " is too big, use the arithmetic mean instead." << endl;
+				cout << "- Warning: sigma of CH " << i + 1 << "_" << j + 1
+					 << " is too big, use the arithmetic mean instead." << endl;
 				mean[i][j] = mean_0[i][j];
 				sigma[i][j] = sigma_0[i][j];
 			}
 		}
 	}
-	eventIter_->ped_trigg_restart();
 	done_flag_ = true;
 }
 
+void PedMeanCalc::close() {
+	if (ped_file_ == NULL)
+		return;
+	for (int i = 0; i < 25; i++) {
+		for (int j = 0; j < 64; j++) {
+			if (mode_ == 'r') {
+				delete f_gaus_[i][j];
+				f_gaus_[i][j] = NULL;
+			}
+			if (mode_ == 'w')
+				h_ped_[i][j]->Write();
+			delete h_ped_[i][j];
+			h_ped_[i][j] = NULL;
+		}
+	}
+	ped_file_->Close();
+	delete ped_file_;
+	ped_file_ = NULL;
+}
+
 void PedMeanCalc::show(int ct_num) {
+	if (ped_file_ == NULL)
+		return;
 	if (!done_flag_)
 		return;
 	int idx = ct_num - 1;
@@ -211,6 +243,8 @@ void PedMeanCalc::print(bool sigma_flag) {
 }
 
 void PedMeanCalc::show_mean() {
+	if (ped_file_ == NULL)
+		return;
 	if (!done_flag_)
 		return;
 	if (h_ped_map_ != NULL) {
@@ -249,7 +283,10 @@ void PedMeanCalc::show_mean() {
 }
 
 void PedMeanCalc::CloseWindow() {
+	if (ped_file_ == NULL)
+		return;
 	cout << "Quitting by user request." << endl;
+	close();
 	gApplication->Terminate(0);
 }
